# docs/swagger.yaml
basePath: /api
definitions:
...
  models.NullInt64:
    properties:
      int64:
        format: int64
        type: integer
      valid:
        description: Valid is true if Int64 is not NULL
        type: boolean
    type: object
  models.NullString:
    properties:
      string:
        type: string
      valid:
        description: Valid is true if String is not NULL
        type: boolean
    type: object
...
  models.Proposal:
    properties:
      client_id:
        type: integer
      client_name:
        type: string
      created_at:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/models.ProposalItem'
        type: array
      notes:
        type: string
      reference:
        type: string
      section_id:
        $ref: '#/definitions/models.NullInt64'
      section_name:
        $ref: '#/definitions/models.NullString'
      status:
        type: string
      updated_at:
        type: string
    type: object
  models.ProposalItem:
    properties:
      id:
        type: integer
      notes:
        type: string
      product_id:
        type: integer
      product_name:
        type: string
      proposal_id:
        type: integer
      quantity:
        type: number
      unit:
        type: string
      unit_price:
        type: number
    type: object
  models.ProposalItemCreate:
    properties:
      notes:
        type: string
      product_id:
        type: integer
      quantity:
        type: number
      unit_price:
        type: number
    required:
    - product_id
    - quantity
    - unit_price
    type: object
  models.ProposalItemUpdate:
    properties:
      notes:
        type: string
      product_id:
        type: integer
      quantity:
        type: number
      unit_price:
        type: number
    required:
    - product_id
    - quantity
    - unit_price
    type: object
  models.Section:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  models.SectionCreate:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  models.SectionSearchResult:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.SectionUpdate:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
host: localhost:5000
info:
  contact:
    email: support@aura-erp.com
    name: API Support
    url: http://www.aura-erp.com/support
  description: This is a powerful ERP backend for Aura.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Aura ERP API
  version: "1.0"
paths:
...
  /proposal-items:
    get:
      consumes:
      - application/json
      description: Get all items associated with a specific proposal
      parameters:
      - description: Proposal ID
        in: query
        name: proposal_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ProposalItem'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List proposal items
      tags:
      - proposals
    post:
      consumes:
      - application/json
      description: Add a new item to a proposal
      parameters:
      - description: Proposal ID
        in: query
        name: proposal_id
        required: true
        type: integer
      - description: Proposal item details
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.ProposalItemCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ProposalItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a proposal item
      tags:
      - proposals
  /proposal-items/{id}:
    delete:
      consumes:
      - application/json
      description: Remove an item from a proposal
      parameters:
      - description: Proposal Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a proposal item
      tags:
      - proposals
    put:
      consumes:
      - application/json
      description: Update details of an existing proposal item
      parameters:
      - description: Proposal Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated proposal item details
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.ProposalItemUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProposalItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a proposal item
      tags:
      - proposals
  /proposals:
    get:
      consumes:
      - application/json
      description: Get all proposals in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Proposal'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List proposals
      tags:
      - proposals
    post:
      consumes:
      - application/json
      description: Create a new proposal
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Proposal'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a proposal
      tags:
      - proposals
  /proposals/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a proposal from the system
      parameters:
      - description: Proposal ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a proposal
      tags:
      - proposals
    get:
      consumes:
      - application/json
      description: Get proposal by ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Proposal'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get proposal by ID
      tags:
      - proposals
    put:
      consumes:
      - application/json
      description: Update a proposal
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Proposal'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a proposal
      tags:
      - proposals
  /proposals/search:
    get:
      consumes:
      - application/json
      description: Get all proposals in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Proposal'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List proposals
      tags:
      - proposals
  /sections:
    get:
      consumes:
      - application/json
      description: Get all sections in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Section'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List sections
      tags:
      - sections
    post:
      consumes:
      - application/json
      description: Create a new section
      parameters:
      - description: Section details
        in: body
        name: section
        required: true
        schema:
          $ref: '#/definitions/models.SectionCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Section'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a section
      tags:
      - sections
  /sections/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a section from the system
      parameters:
      - description: Section ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a section
      tags:
      - sections
    get:
      consumes:
      - application/json
      description: Get a single section by its ID
      parameters:
      - description: Section ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Section'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get section by ID
      tags:
      - sections
    put:
      consumes:
      - application/json
      description: Update an existing section's details
      parameters:
      - description: Section ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated section details
        in: body
        name: section
        required: true
        schema:
          $ref: '#/definitions/models.SectionUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Section'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a section
      tags:
      - sections
  /sections/search:
    get:
      consumes:
      - application/json
      description: Search sections by name
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - default: 25
        description: Limit results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SectionSearchResult'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search sections
      tags:
      - sections
...
swagger: "2.0"